# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JobBit APIs",
    "version": "v1"
  },
  "servers": [
    { }
  ],
  "paths": {
...
    "/api/Companies/GetAllCompanies": {
      "get": {
        "tags": [
          "Companies"
        ],
        "operationId": "GetAllCompanies",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyListDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyListDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyListDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Companies/GetAllApplicantsForCompanyJob/{CompanyID}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "operationId": "GetAllApplicantsForCompanyJob",
        "parameters": [
          {
            "name": "CompanyID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListApplicantForCompanyJobCaterizedByStatus"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListApplicantForCompanyJobCaterizedByStatus"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListApplicantForCompanyJobCaterizedByStatus"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Companies/GetCompanyByID/{CompanyID}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "operationId": "GetCompanyByID",
        "parameters": [
          {
            "name": "CompanyID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AllCompanyInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllCompanyInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllCompanyInfo"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Companies/GetCompanyByUserID/{UserID}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "operationId": "GetCompanyByUserID",
        "parameters": [
          {
            "name": "UserID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AllCompanyInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllCompanyInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllCompanyInfo"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Companies/GetCompanyByEmail/{Email}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "operationId": "GetCompanyByEmail",
        "parameters": [
          {
            "name": "Email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AllCompanyInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllCompanyInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllCompanyInfo"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Companies/IsCompanyExistByID/{CompanyID}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "operationId": "IsCompanyExistByID",
        "parameters": [
          {
            "name": "CompanyID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicantForCompanyJobDTO": {
        "type": "object",
        "properties": {
          "requestID": {
            "type": "integer",
            "format": "int32"
          },
          "postedDate": {
            "type": "string",
            "format": "date-time"
          },
          "jobTitle": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CompanyListDTO": {
        "type": "object",
        "properties": {
          "companyID": {
            "type": "integer",
            "format": "int32"
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "wilayaName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ListApplicantForCompanyJobCaterizedByStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "nullable": true
          },
          "statusAsText": {
            "type": "string",
            "nullable": true
          },
          "applicantForCompanyJob": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApplicantForCompanyJobDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}